refactor: migrate to ts
