+++ package.json
@@ -3,30 +3,32 @@
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
   "main": "index.js",
-  "type": "module",
   "scripts": {
     "start": "node ./index.js",
-    "start:dev": "nodemon ./index.js",
+    "start:dev": "nodemon ./src/index.ts",
     "build": "webpack --config webpack.config.mjs",
     "start:prod": "npm run build && node dist/bundle.cjs",
     "lint": "eslint src/**/*.js",
     "lint:fix": "eslint src/**/*.ts --fix",
-    "format": "prettier . --write",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "format": "prettier . --write"
   },
   "author": "Andrei Auchynnikau",
   "keywords": [],
   "license": "ISC",
   "devDependencies": {
+    "@types/node": "^22.8.1",
+    "@types/ws": "^8.5.12",
     "@eslint/js": "^9.13.0",
     "@types/eslint__js": "^8.42.3",
-    "@types/node": "^22.8.1",
     "eslint": "^9.13.0",
-    "nodemon": "^2.0.22",
+    "husky": "^9.0.11",
+    "lint-staged": "^15.2.5",
+    "nodemon": "^3.1.7",
     "prettier": "^3.3.3",
+    "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.6.3",
-    "typescript-eslint": "^8.11.0",
+    "typescript-eslint": "^8.10.0",
     "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4"
   },
