--- src/handlers/player-handler.js
@@ -1,10 +1,0 @@
-export const playerHandler = async (command) => {
-  switch (command) {
-    case 'play':
-      console.log('reg');
-      break;
-
-    default:
-      console.log('unknown command');
-  }
-};
