+++ src/handlers/registration-handlers.ts
@@ -1,0 +1,24 @@
+import { clients, players } from '../state';
+import { Command, FIXME } from '../types';
+import {
+  notifyClientsAboutRoomUpdates,
+  notifyClientsAboutWinnersUpdates,
+  sendSocketMessage,
+} from '../utils';
+
+export const registerPlayerHandler = (ws: FIXME, socketId: string, payload: FIXME) => {
+  const parsedData = JSON.parse(payload.data);
+  const { name, password } = parsedData;
+  const client = clients.get(socketId);
+
+  players.set(socketId, { name, password });
+  const responseData = {
+    name,
+    index: socketId,
+    error: false,
+    errorText: '',
+  };
+  sendSocketMessage(ws, Command.REG, responseData);
+  notifyClientsAboutRoomUpdates();
+  notifyClientsAboutWinnersUpdates([client]);
+};
